Made source code refactoring.
