+++ receiver/googlecloudspannerreceiver/internal/metadata/metadata_test.go
@@ -1,0 +1,35 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package metadata
+
+import "go.opentelemetry.io/collector/model/pdata"
+
+const (
+	labelName       = "LabelName"
+	labelColumnName = "LabelColumnName"
+
+	stringValue  = "stringValue"
+	int64Value   = int64(64)
+	float64Value = float64(64.64)
+	boolValue    = true
+
+	metricName       = "metricName"
+	metricColumnName = "metricColumnName"
+	metricDataType   = pdata.MetricDataTypeGauge
+	metricUnit       = "metricUnit"
+	metricNamePrefix = "metricNamePrefix-"
+
+	timestampColumnName = "INTERVAL_END"
+)
